[VCF-1] Updating schema/parser to be flexible
Updating the variant schema and parser to be more flexible -- the
high-level changes here involve removing the MULT_VAL array, and storing
a lot of the other fields in a generic blob/string field which can be
parsed out later.

The initial change here removes the KGENOMES_MULT_VAL array from the
kgenomes schema file.