Adding a Google Test-based unit testing skeleton.
GTest is the Google C++ Unit Testing framework,
https://code.google.com/p/googletest/

This commit includes a skeleton .cpp file for writing tests, as well as
the Makefile updates to compile an executable to run the tests
(vcfstreamer_test).

Compiling now requires having the GTEST repository
cloned somewhere on the current machine, in order to build libgtest.a --
on salty, this is installed in /home/bio/tdanford/googletest but we
should make this a build parameter in the future.