Adding initial scripts, schemas for GVCF loading.
These are the basic scripts I've written to do GVCF loading.

To load a GVCF, you run the loader script,
  ./loader.sh PATH

where PATH is the path to an (unzipped) GVCF file.

Right now, the sample name is hard-coded, which should be fixed in the future.

Another shortcoming is that this script rebuilds the gvcf_chrom and gvcf_sample
tables completely with each load -- but this should be rewritten as an append of
new values.

It leaves two tmp tables around: 'gvcf_tmp' and 'tmp2'.